Project moved to Visual Studion 2017
Moved project over from visual studio 2013 to 2017. Added a readme in
root folder for changes I had to make for it to run.

+++ Source/Framework/button.cpp
@@ -58,7 +58,7 @@ void Button::SetCustomSprite(Sprite* sprite)
 	m_customSprite = true;
 }
 
-void Button::SetBackgroundColour(int r, int g, int b)
+void Button::SetBackgroundColour(Uint8 r, Uint8 g, Uint8 b)
 {
 	m_backgroundColour = { r, g, b };
 }
